feat: 대시보드 레이아웃 개선 및 랭크 관련 컴포넌트 추가
