--- app/(afterLogin)/(dashboard)/rank/layout.tsx
@@ -1,7 +1,0 @@
-export default function RankLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
